Struct pcg::PcgRng

pub struct PcgRng {
    // some fields omitted
}

A PCG-based random number generator.

The PCG algorithm is not suitable for cryptographic purposes but provides an excellent combination of speed and unpredictability. It is only slightly slower than rand::XorShiftRng but provides much higher-quality output.

This particular implementation uses a 128-bit state value, has a period of 264, and uses the XSH-RR output function.

Methods

impl PcgRng

fn new_unseeded() -> PcgRng

Returns a new PcgRng instance which is not seeded.

The initial values of this RNG are constants, so all generators created by this function will yield the same stream of random numbers. It is highly recommended that this is created through SeedableRng instead of this function.